Rename 'codedStream' folder to 'stream'
Also add a SWIG interface for this directory.
