+++ src/AvTranscoder/avTranscoder.i
@@ -20,12 +20,6 @@
 #include <AvTranscoder/codec/AudioCodec.hpp>
 #include <AvTranscoder/codec/DataCodec.hpp>
 
-#include <AvTranscoder/codedStream/IOutputStream.hpp>
-#include <AvTranscoder/codedStream/OutputStream.hpp>
-
-#include <AvTranscoder/codedStream/IInputStream.hpp>
-#include <AvTranscoder/codedStream/InputStream.hpp>
-
 #include <AvTranscoder/transcoder/StreamTranscoder.hpp>
 #include <AvTranscoder/transcoder/Transcoder.hpp>
 %}
@@ -48,12 +42,7 @@ namespace std {
 %include <AvTranscoder/codec/AudioCodec.hpp>
 %include <AvTranscoder/codec/DataCodec.hpp>
 
-%include <AvTranscoder/codedStream/IOutputStream.hpp>
-%include <AvTranscoder/codedStream/OutputStream.hpp>
-
-%include <AvTranscoder/codedStream/IInputStream.hpp>
-%include <AvTranscoder/codedStream/InputStream.hpp>
-
+%include "AvTranscoder/stream/stream.i"
 %include "AvTranscoder/decoder/decoder.i"
 %include "AvTranscoder/encoder/encoder.i"
 %include "AvTranscoder/file/file.i"
