+++ src/AvTranscoder/stream/IInputStream.hpp
