+++ src/AvTranscoder/stream/InputStream.hpp
@@ -1,5 +1,5 @@
-#ifndef _AV_TRANSCODER_CODED_STREAM_AV_INPUT_STREAM_HPP_
-#define _AV_TRANSCODER_CODED_STREAM_AV_INPUT_STREAM_HPP_
+#ifndef _AV_TRANSCODER_STREAM_INPUT_STREAM_HPP_
+#define _AV_TRANSCODER_STREAM_INPUT_STREAM_HPP_
 
 #include "IInputStream.hpp"
