+++ src/AvTranscoder/stream/stream.i
@@ -1,0 +1,13 @@
+%{
+#include <AvTranscoder/stream/IOutputStream.hpp>
+#include <AvTranscoder/stream/OutputStream.hpp>
+
+#include <AvTranscoder/stream/IInputStream.hpp>
+#include <AvTranscoder/stream/InputStream.hpp>
+%}
+
+%include <AvTranscoder/stream/IOutputStream.hpp>
+%include <AvTranscoder/stream/OutputStream.hpp>
+
+%include <AvTranscoder/stream/IInputStream.hpp>
+%include <AvTranscoder/stream/InputStream.hpp>
